backend/requirements.txt
fastapi==0.110.0
requests==2.31.0
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic==2.6.2
pydantic-settings==2.2.1
pymongo[srv]==4.6.1
dnspython==2.5.0
